Update MoMo payment integration and environment configuration
- Added MoMo payment record creation and status handling in MoMoController.
- Enhanced error handling and logging for payment processing.
- Updated JWT configuration to cast environment variables to integers.
- Modified .env.example with new MoMo API keys and database settings.
- Updated composer.lock to reflect new package versions and dependencies.
- Improved frontend MomoQR component to handle payment status checks and responses.
- Added new API routes for MoMo webhook and status checking.

+++ config/jwt.php
@@ -101,7 +101,8 @@ return [
     |
     */
 
-    'ttl' => env('JWT_TTL', 60),
+    // Cast to int to satisfy Carbon (expects int|float, not string from env)
+    'ttl' => (int) env('JWT_TTL', 60),
 
     /*
     |--------------------------------------------------------------------------
@@ -120,7 +121,7 @@ return [
     |
     */
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160),
 
     /*
     |--------------------------------------------------------------------------
@@ -205,7 +206,7 @@ return [
     |
     */
 
-    'leeway' => env('JWT_LEEWAY', 0),
+    'leeway' => (int) env('JWT_LEEWAY', 0),
 
     /*
     |--------------------------------------------------------------------------
@@ -232,7 +233,7 @@ return [
     |
     */
 
-    'blacklist_grace_period' => env('JWT_BLACKLIST_GRACE_PERIOD', 0),
+    'blacklist_grace_period' => (int) env('JWT_BLACKLIST_GRACE_PERIOD', 0),
 
     /*
     |--------------------------------------------------------------------------
